created home page, add benefit page, split benefit.urls into api and web urls
