+++ benefits/api_urls.py
@@ -1,4 +1,4 @@
-from .views import BenefitView, CategoryViewSet
+from .api_views import BenefitView, CategoryViewSet
 from rest_framework.routers import DefaultRouter
 
 router = DefaultRouter()
