+++ benefits/web_urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .web_views import new_benefit
+
+urlpatterns = [
+    path("new", new_benefit)
+]
