+++ templates/users/create_user_success.html
@@ -1,14 +1,9 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
+{% extends "base.html" %}
+{% block title %}Congrats{% endblock %}
+{% block body %}
     Congrats {{ email }}, you've become a maintainer for Student Benefits,
     <ul>
         <li>Now all your contributions are automatically approved</li>
         <li>You can approve/reject other contributions</li>
     </ul>
-</body>
-</html>
+{% endblock %}
